fix: rename IconPageArrows
